move jit to paddle
