+++ python/setup.py.in
@@ -154,7 +154,7 @@ packages=['paddle',
           'paddle.fleet.proto',
           'paddle.fleet.utils',
           'paddle.framework',
-          'paddle.framework.jit',
+          'paddle.jit',
           'paddle.fluid',
           'paddle.fluid.dygraph',
           'paddle.fluid.dygraph.dygraph_to_static',
